+++ ruoyi-admin/src/main/java/com/ruoyi/web/controller/gauge/PsyGaugeMultiSettingController.java
@@ -64,7 +64,7 @@ public class PsyGaugeMultiSettingController extends BaseController
      */
     @PreAuthorize("@ss.hasPermi('gauge:multiSetting:query')")
     @GetMapping(value = "/{id}")
-    public AjaxResult getInfo(@PathVariable("id") Long id)
+    public AjaxResult getInfo(@PathVariable("id") Integer id)
     {
         return AjaxResult.success(psyGaugeMultiSettingService.selectPsyGaugeMultiSettingById(id));
     }
@@ -97,7 +97,7 @@ public class PsyGaugeMultiSettingController extends BaseController
     @PreAuthorize("@ss.hasPermi('gauge:multiSetting:remove')")
     @Log(title = "心理测评多维设置", businessType = BusinessType.DELETE)
 	@DeleteMapping("/{ids}")
-    public AjaxResult remove(@PathVariable Long[] ids)
+    public AjaxResult remove(@PathVariable Integer[] ids)
     {
         return toAjax(psyGaugeMultiSettingService.deletePsyGaugeMultiSettingByIds(ids));
     }
