+++ ruoyi-paychology/src/main/java/com/ruoyi/gauge/domain/PsyOrderPay.java
@@ -28,11 +28,11 @@ public class PsyOrderPay extends BaseEntity
     private static final long serialVersionUID = 1L;
 
     /** 主键 */
-    private Long id;
+    private Integer id;
 
     /** 订单号 */
     @Excel(name = "订单号")
-    private Long orderId;
+    private Integer orderId;
 
     /** 支付状态(1-待支付,2-已支付，3-取消支付) */
     @Excel(name = "支付状态(1-待支付,2-已支付，3-取消支付)")
