+++ ruoyi-paychology/src/main/java/com/ruoyi/gauge/mapper/PsyConsultingServiceMapper.java
@@ -18,7 +18,7 @@ public interface PsyConsultingServiceMapper
      * @param id 咨询服务主键
      * @return 咨询服务
      */
-    public PsyConsultingService selectPsyConsultingServiceById(Long id);
+    public PsyConsultingService selectPsyConsultingServiceById(Integer id);
 
     /**
      * 查询咨询服务列表
@@ -50,7 +50,7 @@ public interface PsyConsultingServiceMapper
      * @param id 咨询服务主键
      * @return 结果
      */
-    public int deletePsyConsultingServiceById(Long id);
+    public int deletePsyConsultingServiceById(Integer id);
 
     /**
      * 批量删除咨询服务
@@ -58,5 +58,5 @@ public interface PsyConsultingServiceMapper
      * @param ids 需要删除的数据主键集合
      * @return 结果
      */
-    public int deletePsyConsultingServiceByIds(Long[] ids);
+    public int deletePsyConsultingServiceByIds(Integer[] ids);
 }
