+++ ruoyi-paychology/src/main/java/com/ruoyi/gauge/mapper/PsyGaugeMapper.java
@@ -17,7 +17,7 @@ public interface PsyGaugeMapper
      * @param id 心理测评主键
      * @return 心理测评
      */
-    public PsyGauge selectPsyGaugeById(Long id);
+    public PsyGauge selectPsyGaugeById(Integer id);
 
     /**
      * 查询心理测评列表
@@ -49,7 +49,7 @@ public interface PsyGaugeMapper
      * @param id 心理测评主键
      * @return 结果
      */
-    public int deletePsyGaugeById(Long id);
+    public int deletePsyGaugeById(Integer id);
 
     /**
      * 批量删除心理测评
@@ -57,5 +57,5 @@ public interface PsyGaugeMapper
      * @param ids 需要删除的数据主键集合
      * @return 结果
      */
-    public int deletePsyGaugeByIds(Long[] ids);
+    public int deletePsyGaugeByIds(Integer[] ids);
 }
