+++ ruoyi-paychology/src/main/java/com/ruoyi/gauge/mapper/PsyOrderPayMapper.java
@@ -17,7 +17,7 @@ public interface PsyOrderPayMapper
      * @param id 心理咨询订单支付信息主键
      * @return 心理咨询订单支付信息
      */
-    public PsyOrderPay selectPsyOrderPayById(Long id);
+    public PsyOrderPay selectPsyOrderPayById(Integer id);
 
     /**
      * 查询心理咨询订单支付信息列表
@@ -57,7 +57,7 @@ public interface PsyOrderPayMapper
      * @param id 心理咨询订单支付信息主键
      * @return 结果
      */
-    public int deletePsyOrderPayById(Long id);
+    public int deletePsyOrderPayById(Integer id);
 
     /**
      * 批量删除心理咨询订单支付信息
@@ -65,5 +65,5 @@ public interface PsyOrderPayMapper
      * @param ids 需要删除的数据主键集合
      * @return 结果
      */
-    public int deletePsyOrderPayByIds(Long[] ids);
+    public int deletePsyOrderPayByIds(Integer[] ids);
 }
