+++ ruoyi-paychology/src/main/java/com/ruoyi/gauge/service/IPsyConsultingOrderService.java
@@ -18,7 +18,7 @@ public interface IPsyConsultingOrderService
      * @param id 咨询订单主键
      * @return 咨询订单
      */
-    public PsyConsultingOrder selectPsyConsultingOrderById(Long id);
+    public PsyConsultingOrder selectPsyConsultingOrderById(Integer id);
 
     /**
      * 查询咨询订单列表
@@ -50,7 +50,7 @@ public interface IPsyConsultingOrderService
      * @param ids 需要删除的咨询订单主键集合
      * @return 结果
      */
-    public int deletePsyConsultingOrderByIds(Long[] ids);
+    public int deletePsyConsultingOrderByIds(Integer[] ids);
 
     /**
      * 删除咨询订单信息
@@ -58,5 +58,5 @@ public interface IPsyConsultingOrderService
      * @param id 咨询订单主键
      * @return 结果
      */
-    public int deletePsyConsultingOrderById(Long id);
+    public int deletePsyConsultingOrderById(Integer id);
 }
